refactored junit tests into packages
