+++ docs/UserGuide.md
@@ -86,6 +86,7 @@ The highlights and symbols used in this document are as follow:
 4. You can exit the program by running `bye`.
 
 ##Initialisation
+
 When you first open Plan&Score, you will be greeted by a welcome message as well as 
 a prompt asking for your name. This is as shown below.
