+++ src/test/java/seedu/duke/storage/StorageTest.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.storage;
 
 import org.junit.jupiter.api.Test;
 import seedu.duke.exception.StorageCorruptedException;
